# packages/api-contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mindfolio-api",
    "description": "Mindfolio 管理端 API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/task-completions": {
      "get": {
        "tags": [
          "completions"
        ],
        "description": "完成事实按完成时刻倒序；date 按 Asia/Shanghai 业务日期筛选。",
        "operationId": "list_task_completions",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionPage"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_session": []
          }
        ]
      }
    },
    "/task-completions/days": {
      "get": {
        "tags": [
          "completions"
        ],
        "description": "按 Asia/Shanghai 业务日期汇总完成次数，每页最多 50 天。",
        "operationId": "list_task_completion_days",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayPage"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        }
      },
      "CompletionPage": {
        "type": "object",
        "required": [
          "items",
          "page",
          "has_more"
        ],
        "properties": {
          "has_more": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompletionResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "CompletionResponse": {
        "type": "object",
        "required": [
          "id",
          "task_id",
          "task_title",
          "completed_at",
          "business_date"
        ],
        "properties": {
          "business_date": {
            "type": "string"
          },
          "completed_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "project_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "task_id": {
            "type": "string"
          },
          "task_title": {
            "type": "string"
          }
        }
      },
...
      "DayPage": {
        "type": "object",
        "required": [
          "items",
          "page",
          "has_more"
        ],
        "properties": {
          "has_more": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DaySummary"
            }
          },
          "page": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "DaySummary": {
        "type": "object",
        "required": [
          "business_date",
          "completed_count"
        ],
        "properties": {
          "business_date": {
            "type": "string"
          },
          "completed_count": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "database_unavailable",
          "unauthorized",
          "invalid_credentials",
          "invalid_request",
          "forbidden",
          "rate_limited",
          "auth_unavailable",
          "not_found",
          "version_conflict"
        ]
      },
...
    },
    "securitySchemes": {
      "admin_session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-mf_session"
      }
    }
  }
}